test: run UP Init tests in separate test suite
