+++ package.json
@@ -43,6 +43,7 @@
     "test:parallel": "run-p test:mocks test:up test:lsp1 test:lsp2 test:lsp4 test:lsp6 test:lsp7 test:lsp8 test:lsp9 test:lsp11 test:universalfactory",
     "test:mocks": "hardhat test --no-compile tests/Mocks/*.test.ts ",
     "test:up": "hardhat test --no-compile tests/UniversalProfile.test.ts",
+    "test:upinit": "hardhat test --no-compile tests/UniversalProfileInit.test.ts",
     "test:lsp1": "hardhat test --no-compile tests/LSP1UniversalReceiver/*.test.ts",
     "test:lsp2": "hardhat test --no-compile tests/LSP2ERC725YJSONSchema/LSP2UtilsLibrary.test.ts",
     "test:lsp4": "hardhat test --no-compile tests/LSP4DigitalAssetMetadata/LSP4Compatibility.test.ts",
